Remove CodableKitShared and refactor option types
Eliminates the CodableKitShared target and migrates CodableKeyOptions and CodableOptions into CodableKit and CodableKitMacros, respectively. Updates all references and implementations to use the new locations, and removes unnecessary imports. This simplifies the package structure and clarifies the separation between runtime and macro-related code.
