+++ Sources/CodableKitMacros/CodableMacro.swift
@@ -5,7 +5,6 @@
 //  Created by Wendell on 3/30/24.
 //
 
-import CodableKitShared
 import Foundation
 import SwiftDiagnostics
 import SwiftSyntax
