+++ Sources/CodableKitMacros/CodableOptions.swift
@@ -2,33 +2,24 @@
 //  CodableOptions.swift
 //  CodableKit
 //
-//  Created by Wendell Wang on 2025/1/13.
+//  Created by Wendell Wang on 2025/10/2.
 //
 
 import SwiftSyntax
 
-/// Options that customize the behavior of the `@Codable` macro expansion.
-public struct CodableOptions: OptionSet, Sendable {
-  public let rawValue: Int32
+struct CodableOptions: OptionSet, Sendable {
+  let rawValue: Int32
 
-  public init(rawValue: Int32) {
+  init(rawValue: Int32) {
     self.rawValue = rawValue
   }
 
-  /// The default options, which perform standard Codable expansion with super encode/decode calls.
-  public static let `default`: Self = []
-
-  /// Skips generating super encode/decode calls in the expanded code.
-  ///
-  /// Use this option when the superclass doesn't conform to `Codable`.
-  /// When enabled:
-  /// - Replaces `super.init(from: decoder)` with `super.init()`
-  /// - Removes `super.encode(to: encoder)` call entirely
-  public static let skipSuperCoding = Self(rawValue: 1 << 0)
+  static let `default`: Self = []
+  static let skipSuperCoding = Self(rawValue: 1 << 0)
 }
 
 extension CodableOptions {
-  package init(from expr: MemberAccessExprSyntax) {
+  init(from expr: MemberAccessExprSyntax) {
     let variableName = expr.declName.baseName.text
     switch variableName {
     case "skipSuperCoding":
@@ -42,7 +33,7 @@ extension CodableOptions {
 extension CodableOptions {
   /// Parse the options from 1a `LabelExprSyntax`. It support parse a single element like `.default`,
   /// or multiple elements like `[.ignored, .explicitNil]`
-  package static func parse(from labeledExpr: LabeledExprSyntax) -> Self {
+  static func parse(from labeledExpr: LabeledExprSyntax) -> Self {
     if let memberAccessExpr = labeledExpr.expression.as(MemberAccessExprSyntax.self) {
       Self.init(from: memberAccessExpr)
     } else if let arrayExpr = labeledExpr.expression.as(ArrayExprSyntax.self) {
@@ -61,7 +52,7 @@ extension LabeledExprSyntax {
   /// or multiple elements like [.ignored, .explicitNil].
   ///
   /// This is a convenience method to use for chaining.
-  package func parseCodableOptions() -> CodableOptions {
+  func parseCodableOptions() -> CodableOptions {
     CodableOptions.parse(from: self)
   }
 }
